refactor: 실행 결과, 최종 우승자 출력 함수 파일로 분리
